# src/main/java/com/sprint/mission/discodeit/dto/channel/ChannelResponse.java
package com.sprint.mission.discodeit.dto.channel;

import com.sprint.mission.discodeit.entity.ChannelType;
import java.time.Instant;
import java.util.UUID;

@Deprecated
public record ChannelResponse(
    UUID id,
    Instant createdAt,
    Instant updatedAt,
    ChannelType type,
    String name,
    String description
) {

}