# src/main/java/com/sprint/mission/discodeit/dto/channel/PublicChannelCreateRequest.java
package com.sprint.mission.discodeit.dto.channel;

import jakarta.validation.constraints.NotBlank;

public record PublicChannelCreateRequest(
    @NotBlank(message = "채널명은 필수입니다.")
    String name,

    String description
) {

}